requirements.txt
celery==5.0.5
Django==3.0.5
django-cors-headers==3.4.0
django-mongoengine==0.4.4
django-rest-framework-mongoengine==3.4.1
django-timezone-field==4.1.1
djangorestframework==3.11.0
djangotoolbox==1.8.0
djongo==1.3.3
fields==5.0.0
mongoadmin==0.2
mongodbforms==0.3
mongoengine==0.20.0
pymodm==0.4.3
pytz==2020.1
redis==3.5.3
